Cake Controller to create Excel sheets
Cakephp controller to create and index Excel Sheets based on info from a user. The User can input data onto a table from a mobile device or browser and this controller is in charge of creating a proper excel sheet with the correct logic. It does the logic based on guidelines given by the DEC.